+++ spec/praxis/media_type_spec.rb
@@ -74,6 +74,22 @@ describe Praxis::MediaType do
     end
   end
 
+  context 'with nodoc! option' do
+    subject(:media_type) do
+      Class.new(Praxis::MediaType) do
+        nodoc!
+        attributes do
+          attribute :name, String
+        end
+      end
+    end
+
+    it 'has the :doc_visibility set' do
+      expect(media_type.options[:doc_visibility]).to be(:nodoc)
+    end
+
+  end
+
   context 'using blueprint caching' do
     it 'has specs'
   end